requirements.txt
pillow
python-twitter
requests